Fix selectivity calculation again

+++ src/currentia/core/operator/operator-selection.h
@@ -57,7 +57,7 @@ namespace currentia {
         double get_selectivity() {
             if (input_tuple_count_ == 0)
                 return 0.0;
-            return selected_tuple_count_ / input_tuple_count_;
+            return static_cast<double>(selected_tuple_count_) / static_cast<double>(input_tuple_count_);
         }
     };
 }
